# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Honey School API Gateway
  description: Public contract for the Honey School API Gateway.
  version: 0.1.0
servers:
- url: /api
  description: Honey School web-app API proxy
paths:
...
  /schedule/lessons/{lessonId}/lobby/{attemptId}/deny:
    post:
      tags:
      - lesson-admission-controller
      operationId: deny
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessStatusResponse"
  /schedule/lessons/{lessonId}/lobby/{attemptId}/approve:
    post:
      tags:
      - lesson-admission-controller
      operationId: approve
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonLobbyDecisionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessStatusResponse"
...
  /schedule/lessons/{lessonId}/admissions/{subject}/readmit:
    post:
      tags:
      - lesson-admission-controller
      operationId: readmit
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: subject
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonAdmissionActionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessStatusResponse"
  /schedule/lessons/{lessonId}/admissions/{subject}/kick:
    post:
      tags:
      - lesson-admission-controller
      operationId: kick
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: subject
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonAdmissionActionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessStatusResponse"
  /schedule/lessons/{lessonId}/access-link/rotate:
    post:
      tags:
      - lesson-access-controller
      operationId: rotate
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessLinkResponse"
  /schedule/lessons/{lessonId}/access-attempts/{attemptId}/remembered:
    post:
      tags:
      - lesson-admission-controller
      operationId: remembered
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Honey-Lesson-Attempt
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessAttemptResponse"
...
  /public/lesson-access/{lessonId}/start:
    post:
      tags:
      - lesson-access-controller
      operationId: start
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Origin
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonAccessStartRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessAttemptResponse"
  /public/lesson-access/{lessonId}/attempts/{attemptId}/lobby:
    post:
      tags:
      - lesson-entry-controller
      operationId: requestLobby
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Honey-Lesson-Attempt
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonLobbyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessStatusResponse"
  /public/lesson-access/{lessonId}/attempts/{attemptId}/email-code:
    post:
      tags:
      - lesson-entry-controller
      operationId: requestEmailCode
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Honey-Lesson-Attempt
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonEmailCodeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessStatusResponse"
  /public/lesson-access/{lessonId}/attempts/{attemptId}/email-code/verify:
    post:
      tags:
      - lesson-entry-controller
      operationId: verifyEmailCode
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Honey-Lesson-Attempt
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonEmailCodeVerifyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessAttemptResponse"
...
  /schedule/lessons/{lessonId}/admissions:
    get:
      tags:
      - lesson-admission-controller
      operationId: overview
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAdmissionOverviewResponse"
  /schedule/lessons/{lessonId}/access-link:
    get:
      tags:
      - lesson-access-controller
      operationId: getOrCreate
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessLinkResponse"
    delete:
      tags:
      - lesson-access-controller
      operationId: revoke
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
...
  /public/lesson-access/{lessonId}/attempts/{attemptId}/status:
    get:
      tags:
      - lesson-entry-controller
      operationId: status
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Honey-Lesson-Attempt
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonAccessAttemptResponse"
...
components:
  schemas:
...
    LessonAccessStatusResponse:
      type: object
      properties:
        status:
          type: string
      required:
      - status
    LessonLobbyDecisionRequest:
      type: object
      properties:
        studentSubject:
          type: string
          maxLength: 255
          minLength: 0
        expectedRevision:
          type:
          - integer
          - "null"
          format: int64
      required:
      - studentSubject
...
    LessonAdmissionActionRequest:
      type: object
      properties:
        expectedRevision:
          type:
          - integer
          - "null"
          format: int64
    LessonAccessLinkResponse:
      type: object
      properties:
        lessonId:
          type: string
          format: uuid
        url:
          type: string
        revision:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        revokedAt:
          type:
          - string
          - "null"
          format: date-time
      required:
      - createdAt
      - lessonId
      - revision
      - url
    LessonAccessAttemptResponse:
      type: object
      properties:
        attemptId:
          type: string
          format: uuid
        attemptSecret:
          type:
          - string
          - "null"
        status:
          type: string
        lessonId:
          type:
          - string
          - "null"
          format: uuid
        opensAt:
          type:
          - string
          - "null"
          format: date-time
        retryAfterSeconds:
          type:
          - integer
          - "null"
          format: int64
        authorizationUrl:
          type:
          - string
          - "null"
      required:
      - attemptId
      - status
...
    LessonAccessStartRequest:
      type: object
      properties:
        token:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - token
    LessonLobbyRequest:
      type: object
      properties:
        displayLabel:
          type: string
          maxLength: 120
          minLength: 0
      required:
      - displayLabel
    LessonEmailCodeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          minLength: 0
        locale:
          type:
          - string
          - "null"
          maxLength: 16
          minLength: 0
      required:
      - email
    LessonEmailCodeVerifyRequest:
      type: object
      properties:
        code:
          type: string
          maxLength: 12
          minLength: 6
        rememberMe:
          type: boolean
      required:
      - code
      - rememberMe
...
    LessonAdmissionOverviewResponse:
      type: object
      properties:
        lessonId:
          type: string
          format: uuid
        pendingLobby:
          type: array
          items:
            $ref: "#/components/schemas/LessonLobbyEntryResponse"
        admissions:
          type: array
          items:
            $ref: "#/components/schemas/LessonAdmissionResponse"
      required:
      - admissions
      - lessonId
      - pendingLobby
    LessonAdmissionResponse:
      type: object
      properties:
        subject:
          type: string
        status:
          type: string
        revision:
          type: integer
          format: int64
        admissionMethod:
          type:
          - string
          - "null"
        updatedAt:
          type: string
          format: date-time
      required:
      - revision
      - status
      - subject
      - updatedAt
    LessonLobbyEntryResponse:
      type: object
      properties:
        attemptId:
          type: string
          format: uuid
        displayLabel:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required:
      - attemptId
      - createdAt
      - displayLabel
      - expiresAt
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT